Moves the instantiation of Cronet engine upstream.
Per @JensenPaul, instances of CronetEngine require a lot of resources. Creating
them is slow and expensive. For this reason, it's recommended to move the
creation of the engine up in the scope of the app to a longer-lived class.

This commit moves the creation of the engine to the activity level. The activity
also reuses a single instance of the class.

This commit also bumps the Cronet dependency to version 70.3538.110.